# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InterestTimelineBucket": {
        "description": "タイムラインの 1 週バケット。",
        "properties": {
          "interest_article_count": {
            "title": "Interest Article Count",
            "type": "integer"
          },
          "topics": {
            "items": {
              "$ref": "#/components/schemas/InterestTimelineTopicStats"
            },
            "title": "Topics",
            "type": "array"
          },
          "week_start": {
            "format": "date-time",
            "title": "Week Start",
            "type": "string"
          }
        },
        "required": [
          "week_start",
          "interest_article_count",
          "topics"
        ],
        "title": "InterestTimelineBucket",
        "type": "object"
      },
      "InterestTimelineResponse": {
        "description": "`GET /api/interests/timeline` のレスポンス。",
        "properties": {
          "buckets": {
            "items": {
              "$ref": "#/components/schemas/InterestTimelineBucket"
            },
            "title": "Buckets",
            "type": "array"
          }
        },
        "required": [
          "buckets"
        ],
        "title": "InterestTimelineResponse",
        "type": "object"
      },
      "InterestTimelineTopicStats": {
        "description": "タイムラインの 1 週バケット内、1 トピックぶんの集計。",
        "properties": {
          "negative_count": {
            "title": "Negative Count",
            "type": "integer"
          },
          "positive_count": {
            "title": "Positive Count",
            "type": "integer"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "topic",
          "positive_count",
          "negative_count"
        ],
        "title": "InterestTimelineTopicStats",
        "type": "object"
      },
      "InterestTopicItem": {
        "description": "トピック単位の関心 1 件（`user_topic_preferences` 1 行）のレスポンス。",
        "properties": {
          "effective_weight": {
            "title": "Effective Weight",
            "type": "number"
          },
          "negative_weight": {
            "title": "Negative Weight",
            "type": "number"
          },
          "positive_weight": {
            "title": "Positive Weight",
            "type": "number"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "topic",
          "positive_weight",
          "negative_weight",
          "effective_weight",
          "updated_at"
        ],
        "title": "InterestTopicItem",
        "type": "object"
      },
      "InterestTopicListResponse": {
        "description": "`GET /api/interests` のレスポンス。",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InterestTopicItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "InterestTopicListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "TechRadar API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/articles/{article_id}/interest": {
      "delete": {
        "description": "記事を関心記事一覧から除外する。\n\n`user_articles` の該当行を削除するだけで、`article_feedback` には触れない\n（Bad フィードバックも付けない）。「興味がない」（Bad）と「間違って登録した」\n（一覧からの除外）は別の意思表示のため区別する（Issue #14 ヒアリング済み決定事項）。",
        "operationId": "delete_interest_article_api_articles__article_id__interest_delete",
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Article Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Interest Article",
        "tags": [
          "articles"
        ]
      }
    },
...
    "/api/interests": {
      "get": {
        "description": "トピック単位の関心一覧を返す（`PROJECT_SPEC.md` §7.1, §7.2）。\n\n`effective_weight` 降順（同値は `topic` 昇順）で返す。並び順を安定させる\nためにタイブレークを必ず付ける（`recommendation/service.py` の他の並び順と\n同じ方針）。件数はトピックの語彙数に比例し高々数百件程度に収まる見込み\nのため、`cursor` ではなく単純な `offset`/`limit` にする（`api/articles.py`\nの関心記事一覧のような無限スクロール想定の cursor ページングは過剰）。",
        "operationId": "list_interest_topics_api_interests_get",
        "parameters": [
          {
            "description": "取得件数",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "取得件数",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "スキップする件数",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "スキップする件数",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterestTopicListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Interest Topics",
        "tags": [
          "interests"
        ]
      }
    },
...
    "/api/interests/timeline": {
      "get": {
        "description": "関心の推移を週単位のバケットへ集計して返す。\n\n履歴テーブルは新設せず、`user_articles.created_at` と\n`article_feedback.created_at` を集計元にする。`weeks` は `now` から\n`weeks * 7日` 遡った時刻以降を対象にする目安であり、境界週は部分週に\nなりうる（ちょうど週の途中から `since` が始まるため）。\n\nデータが1件も無い週はバケットを作らない（0 件のバケットを並べても\n情報量が無いため）。トピック別集計とは独立に、関心記事の追加件数\n（`user_articles`）も同じ週バケットへ載せて返す。",
        "operationId": "get_interest_timeline_api_interests_timeline_get",
        "parameters": [
          {
            "description": "遡って集計する週数",
            "in": "query",
            "name": "weeks",
            "required": false,
            "schema": {
              "default": 12,
              "description": "遡って集計する週数",
              "maximum": 52,
              "minimum": 1,
              "title": "Weeks",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterestTimelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Interest Timeline",
        "tags": [
          "interests"
        ]
      }
    },
...
  }
}